+++ src/main/java/com/joaocarlos/seplag/Service/UnidadeService.java
@@ -17,7 +17,7 @@ public class UnidadeService {
         return unidadeRepository.findAll();
     }
 
-    public Optional<Unidade> findById(Long id) {
+    public Optional<Unidade> findById(Integer id) {
         return unidadeRepository.findById(id);
     }
 
@@ -25,7 +25,7 @@ public class UnidadeService {
         return unidadeRepository.save(unidade);
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Integer id) {
         unidadeRepository.deleteById(id);
     }
 }
