+++ src/main/java/com/joaocarlos/seplag/repositories/FotoPessoaRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface FotoPessoaRepository extends JpaRepository<FotoPessoa,Long> {
+public interface FotoPessoaRepository extends JpaRepository<FotoPessoa,Integer> {
 }
